rename 컨벤션 정리
