+++ Django_ws/graduation/models.py
@@ -50,7 +50,7 @@ class MyDoneLecture(models.Model):
         return f"{self.lecture_name} ({self.lecture_topic}) ({self.credit})"
     
 #교양 요건 연도별 기준표
-class liberRequire(models.Model):
+class GEStandard(models.Model):
     liber_id = models.AutoField(primary_key=True)
     연도 = models.CharField(max_length=5, blank=True, null=True)
     인간학 = models.DecimalField(max_digits=3, blank=True, null=True, decimal_places=1)
