+++ Django_ws/user/models.py
@@ -12,13 +12,13 @@ class User(models.Model):
     sub_major = models.CharField(max_length=100, blank=True, null=True)
     MD = models.CharField(max_length=100, blank=True, null=True)
     lack_major = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
-    need_general = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
+    lack_GE = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
     lack_sub_major = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
     lack_MD = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
     need_rest = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
     done_major = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
-    done_general = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
-    done_general_rest = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
+    done_GE = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
+    done_GE_rest = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
     done_major_rest = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
     done_rest = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
     done_sub_major = models.DecimalField(max_digits=5, decimal_places=1, blank=True, null=True)
